# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FAD Biometrics Configuration & Results Console — API interna
  description: >
    API interna de la consola (`/api/*`). No debe confundirse con la API externa de FAD
    Biometrics By Steps, documentada en `docs/api-contracts.md`. Toda llamada a FAD pasa por el
    backend a través de `FadApiAdapter`; el frontend nunca contacta a FAD directamente.
  version: "0.1.0"
servers:
  - url: /api
security:
  - cookieAuth: []
...
paths:
...
  /auth/bootstrap/admin:
    post:
      tags: [auth]
      summary: Crea el primer usuario ADMIN (solo si aún no existe ningún usuario)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
      responses:
        "201": { description: Administrador creado, sesión iniciada }
        "409": { description: Ya existe al menos un usuario }
  /auth/login:
    post:
      tags: [auth]
      summary: Inicia sesión (limitado a 20 intentos / 15 min por IP)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: Sesión iniciada (cookie httpOnly) }
        "401": { description: Credenciales incorrectas }
...
  /auth/me:
    get:
      tags: [auth]
      summary: Usuario autenticado actual
      responses:
        "200": { description: OK }
        "401": { description: No autenticado }
...
  /environments/{id}:
    get:
      tags: [environments]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: OK } }
    put:
      tags: [environments]
      summary: Actualiza un ambiente (ADMIN). Un campo de credencial vacío conserva el valor existente.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: Actualizado } }
    delete:
      tags: [environments]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "204": { description: Eliminado } }
  /environments/{id}/credentials/{field}:
    delete:
      tags: [environments]
      summary: Elimina una credencial guardada (ADMIN)
      parameters:
        - { $ref: "#/components/parameters/id" }
        - name: field
          in: path
          required: true
          schema:
            type: string
            enum: [basicAuthUsername, basicAuthPassword, apiUsername, apiPassword, webhookUsername, webhookPassword]
      responses: { "200": { description: OK } }
  /environments/{id}/test-connection:
    post:
      tags: [environments]
      summary: Prueba la autenticación OAuth contra FAD (ADMIN). Nunca devuelve el token.
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200":
          description: Resultado de la prueba
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  code: { type: string }
                  tokenType: { type: string }
                  expiresIn: { type: integer }
...
  /providers/{id}:
    put:
      tags: [providers]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: Actualizado } }
    delete:
      tags: [providers]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "204": { description: Eliminado } }
...
  /executions:
    get:
      tags: [executions]
      summary: Historial de validaciones, con filtros (status, environmentId, templateId, search)
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: environmentId, in: query, schema: { type: string } }
        - { name: templateId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses: { "200": { description: OK } }
    post:
      tags: [executions]
      summary: Crea una validación REAL contra FAD (ADMIN, OPERATOR). 400 si el ambiente no tiene credenciales.
      responses:
        "201": { description: Creada }
        "400": { description: Ambiente sin credenciales configuradas }
...
  /executions/{id}:
    get:
      tags: [executions]
      summary: Detalle de una validación (pasos, webhooks, JSON normalizado y original)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: OK } }
  /executions/{id}/sync:
    post:
      tags: [executions]
      summary: Vuelve a consultar getValidationStep/getValidationData y actualiza el estado normalizado
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses: { "200": { description: Actualizado } }
  /executions/{id}/reveal/{field}:
    post:
      tags: [executions]
      summary: Revela temporalmente key o vector (ADMIN, OPERATOR); queda auditado
      parameters:
        - { $ref: "#/components/parameters/id" }
        - name: field
          in: path
          required: true
          schema: { type: string, enum: [key, vector] }
      responses: { "200": { description: OK } }
  /executions/{id}/steps/{stepKey}/save:
    post:
      tags: [executions]
      summary: Módulo técnico avanzado — reenvía un payload ya cifrado a saveValidationStep
      parameters:
        - { $ref: "#/components/parameters/id" }
        - name: stepKey
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [encryptedPayload]
              properties:
                encryptedPayload: { type: string }
      responses: { "200": { description: OK } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: fad_console_session
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }